code about video 22(The Principle of Least Privilege)

+++ app/Http/Controllers/Api/V1/AuthorTicketsController.php
@@ -10,9 +10,7 @@ use App\Http\Resources\V1\TicketResource;
 use App\Models\Ticket;
 use App\Models\User;
 use App\Policies\V1\TicketPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
-use Illuminate\Http\Request;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 
 class AuthorTicketsController extends ApiController
@@ -28,26 +26,16 @@ class AuthorTicketsController extends ApiController
     }
 
     /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
      * Store a newly created resource in storage.
      */
     public function store(User $author, StoreTicketRequest $request): TicketResource|JsonResponse
     {
-        try {
-            $this->isAble('store', Ticket::class);
+        if ($this->isAble('store', Ticket::class)) {
             return new TicketResource($author->tickets()->create($request->mappedAttributes([
                 'author' => 'user_id',
             ])));
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to create ticket', 403);
         }
+        return $this->error('You are not authorized to create ticket', 403);
     }
 
     /**
@@ -59,31 +47,20 @@ class AuthorTicketsController extends ApiController
     }
 
     /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        //
-    }
-
-    /**
      * Update the specified resource in storage.
      */
     public function update(int $authorId, Ticket $ticket, UpdateTicketRequest $request): JsonResponse|TicketResource
     {
-        try {
-            $this->isAble('update', $ticket);
+        if ($this->isAble('update', $ticket)) {
             $ticket->update($request->mappedAttributes());
             return TicketResource::make($ticket);
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to update this resource', 403);
         }
+        return $this->error('You are not authorized to update this resource', 403);
     }
 
     public function replace(int $authorId, Ticket $ticket, ReplaceTicketRequest $request): JsonResponse|TicketResource
     {
-        try {
-            $this->isAble('replace', $ticket);
+        if ($this->isAble('replace', $ticket)) {
             if ($authorId === $ticket->user_id) {
                 $ticket->update($request->mappedAttributes());
 
@@ -91,9 +68,8 @@ class AuthorTicketsController extends ApiController
             }
             // TODO: ticket doesn't belong to user
             return $this->error('use has no power', 403);
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to edit this resource', 403);
         }
+        return $this->error('You are not authorized to edit this resource', 403);
     }
 
     /**
@@ -101,12 +77,10 @@ class AuthorTicketsController extends ApiController
      */
     public function destroy(int $authorId, Ticket $ticket): JsonResponse
     {
-        try {
-            $this->isAble('delete', $ticket);
+        if ($this->isAble('delete', $ticket)) {
             $ticket->delete();
             return $this->ok('Ticket successfully deleted');
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to delete this resource', 403);
         }
+        return $this->error('You are not authorized to delete this resource', 403);
     }
 }
