+++ app/Http/Controllers/Api/V1/TicketController.php
@@ -9,7 +9,6 @@ use App\Http\Requests\Api\V1\UpdateTicketRequest;
 use App\Http\Resources\V1\TicketResource;
 use App\Models\Ticket;
 use App\Policies\V1\TicketPolicy;
-use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Resources\Json\AnonymousResourceCollection;
 use Illuminate\Support\Facades\Auth;
@@ -31,22 +30,19 @@ class TicketController extends ApiController
      */
     public function store(StoreTicketRequest $request): JsonResponse|TicketResource
     {
-        try {
-            $this->isAble('store', Ticket::class);
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to create ticket', 403);
+        if ($this->isAble('store', Ticket::class)) {
+            $user = Auth::user();
+            // 这里是不是没有必要根据请求参数来获取对应的用户
+            // try {
+            //     $user = User::query()->findOrFail($request->input('data.relationships.author.data.id'));
+            // } catch (ModelNotFoundException $e) {
+            //     return $this->ok('User not found', [
+            //         'error' => 'The provided user id does not exists',
+            //     ]);
+            // }
+            return TicketResource::make($user->tickets()->create($request->mappedAttributes()));
         }
-        $user = Auth::user();
-        // 这里是不是没有必要根据请求参数来获取对应的用户
-        // try {
-        //     $user = User::query()->findOrFail($request->input('data.relationships.author.data.id'));
-        // } catch (ModelNotFoundException $e) {
-        //     return $this->ok('User not found', [
-        //         'error' => 'The provided user id does not exists',
-        //     ]);
-        // }
-
-        return TicketResource::make($user->tickets()->create($request->mappedAttributes()));
+        return $this->error('You are not authorized to create ticket', 403);
     }
 
     /**
@@ -65,24 +61,20 @@ class TicketController extends ApiController
      */
     public function update(Ticket $ticket, UpdateTicketRequest $request): TicketResource|JsonResponse
     {
-        try {
-            $this->isAble('update', $ticket);
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to update that resource', 403);
+        if ($this->isAble('update', $ticket)) {
+            $ticket->update($request->mappedAttributes());
+            return TicketResource::make($ticket);
         }
-        $ticket->update($request->mappedAttributes());
-        return TicketResource::make($ticket);
+        return $this->error('You are not authorized to update that resource', 403);
     }
 
     public function replace(Ticket $ticket, ReplaceTicketRequest $request): TicketResource|JsonResponse
     {
-        try {
-            $this->isAble('replace', $ticket);
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to replace that resource', 403);
+        if ($this->isAble('replace', $ticket)) {
+            $ticket->update($request->mappedAttributes());
+            return TicketResource::make($ticket);
         }
-        $ticket->update($request->mappedAttributes());
-        return TicketResource::make($ticket);
+        return $this->error('You are not authorized to replace that resource', 403);
     }
 
     /**
@@ -90,13 +82,11 @@ class TicketController extends ApiController
      */
     public function destroy(Ticket $ticket): JsonResponse
     {
-        try {
-            $this->isAble('delete', $ticket);
+        if ($this->isAble('delete', $ticket)) {
             // use route model binding will expose some back-end info
             $ticket->delete();
             return $this->ok('Ticket successfully deleted');
-        } catch (AuthorizationException $e) {
-            return $this->error('You are not authorized to delete this resource', 403);
         }
+        return $this->error('You are not authorized to delete this resource', 403);
     }
 }
